requirements.txt
github
PyGithub
openai
requests
langchain_community
tiktoken
python-magic
python-magic-bin;platform_system=="Windows"
nano_graphrag
SentenceTransformer